refactor: rename interfaces package as a functionality package
